requirements.txt
audio_recorder_streamlit==0.0.10
langchain==1.2.10
langchain_community==0.4.1
langchain_core==1.2.17
langchain_google_genai==4.2.1
langchain_huggingface==1.2.1
langchain_pinecone==0.2.13
numpy==2.4.2
pinecone==8.1.0
protobuf==7.34.0
python-dotenv==1.2.2
sentence_transformers==5.2.3
streamlit==1.55.0
streamlit_mic_recorder==0.0.8
torch==2.10.0
